+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/generators/AssistedFactoryImplFirGenerator.kt
@@ -229,7 +229,9 @@ internal class AssistedFactoryImplFirGenerator(session: FirSession) :
           Keys.Default,
           Symbols.Names.create,
           returnTypeProvider = {
-            implClass.source.constructType(it).wrapInProviderIfNecessary(session)
+            implClass.source
+              .constructType(it)
+              .wrapInProviderIfNecessary(session, Symbols.ClassIds.metroProvider)
           },
         ) {
           // Delegate factory