+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/transformers/TopLevelInjectTest.kt
@@ -98,7 +98,8 @@ class TopLevelInjectTest : MetroCompilerTest() {
             }
           """
             .trimIndent()
-        )
+        ),
+        debug = true,
       )
 
     val graph =